oxish: move listener loop into the library
